Feat automated tests - Unit testing

+++ src/users/application/users.service.ts
@@ -11,7 +11,6 @@ import {
 } from './ports/user-repository';
 import { CreateUserDto } from '../http/dto/create-user.dto';
 import { User } from '../domain/model/user';
-import bcrypt from 'bcrypt';
 
 @Injectable()
 export class UsersService implements UserServiceInterface {
@@ -30,14 +29,23 @@ export class UsersService implements UserServiceInterface {
     return await this.create(createUser);
   }
 
-  async create(createUser: CreateUserDto): Promise<User> {
-    const hashedPassword = await bcrypt.hash(createUser.password, 10);
-    const newUser = new User();
-    newUser.setName(createUser.name);
-    newUser.setEmail(createUser.email);
-    newUser.setPassword(hashedPassword);
-    return this.userRepository.registerUser(newUser);
+  async checkPassword(receivedPassword: string, email: string): Promise<boolean> {
+    const persistedPassword = (await this.findUserByEmail(email)).password;
+    return receivedPassword == persistedPassword
   }
+
+  async create(createUser: CreateUserDto): Promise<User> {
+      const newUser = new User(
+        createUser.name,
+        createUser.email,
+        createUser.password,
+        createUser.weight,
+        createUser.yearOfBirth,
+        createUser.city,
+        createUser.state
+      );
+      return this.userRepository.registerUser(newUser);
+  };
 
   private validateAge(user: CreateUserDto) {
     const yearNow = new Date().getFullYear();
@@ -50,11 +58,6 @@ export class UsersService implements UserServiceInterface {
       );
     }
   }
-
-  /* async hashPassword(password: string): Promise<string> {
-    const hashPass = await bcrypt.hash(password, 10);
-    return hashPass;
-  } */
 
   async findAllUsers(): Promise<User[]> {
     return this.userRepository.getAllUsers();
@@ -70,9 +73,12 @@ export class UsersService implements UserServiceInterface {
 
   async modifyWeight(
     email: string,
+    password: string,
     weight: number,
     waterIntake?: number,
   ): Promise<User> {
+    if (!(await this.checkPassword(password, email))) {
+      throw new ForbiddenException('Senha incorreta');}
     const user = await this.findUserByEmail(email);
     user.setWeight(weight);
     user.setWaterIntake(waterIntake);
@@ -80,7 +86,9 @@ export class UsersService implements UserServiceInterface {
     return user;
   }
 
-  async modifyCityAndState(email: string, city: string, state: string): Promise<User> {
+  async modifyCityAndState(email: string, password: string, city: string, state: string): Promise<User> {
+    if (!(await this.checkPassword(password, email))) {
+      throw new ForbiddenException('Senha incorreta');}
     const user = await this.findUserByEmail(email);
     user.setCity(city);
     user.setState(state);
@@ -88,17 +96,10 @@ export class UsersService implements UserServiceInterface {
     return user;
   }
 
-  async modifyCityState(email:string,city: string, state: string):Promise<User>{
-    
-    const user = await this.findUserByEmail(email)
-    user.setCity(city)
-    user.setState(state)
-    
-    this.userRepository.modifySave(user)
-    return user
-  }
-
-  async deleteUser(email: string) {
+  async deleteUser(email: string, password: string): Promise<string> {
+    if (!(await this.checkPassword(password, email))) {
+      throw new ForbiddenException('Senha incorreta');}
     await this.userRepository.deleteUser(email);
+    return 'Usuário deletado com sucesso!';
   }
 }
